Change app name from "ACbus" to "ASEAG"
Left the original author's name intact, credit where credit is due. Changed the name
to make it more meaningful when browsing the apps on the watch, and ASEAG is what it's
all about, isn't it? Hopefully ASEAG will not protest against using their name. Might
have to include a disclaimer somewhere.