Prevent deadlock (never updates again) after offline
Under certain circumstances, the "updating" flag would be
set, the update would fail, and the "updating" flag would
never be reset, so the app would never trigger a new update
again (always thinking that an update was still in progress).
The solution is to do away with the "updating" flag and,
in the worst case, start a new update while there is still
one ongoing.